requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-HTTPAuth==4.2.0
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.1
requests-cache==0.5.2
scikit-image==0.18.1
scipy==1.6.0
six==1.15.0
tifffile==2021.1.11
urllib3==1.26.2
Werkzeug==1.0.1
